上游新增:
- McpAgentServer — agent 能力暴露为 MCP stdio server
- Agent Session 持久化基础设施（session_events 表 + v139 migration）
- per-session cancel token + HarnessAgentAdapter 真实化集成
- Windows 真沙箱 + ApprovalPolicy 策略化审批
- TokenBudget Compact 自动压缩
- clean_output 上移至 harness，解除 agent→runtime-core 越界依赖
- 删除 coordinator/adapter/config 等冗余模块
- 前端 agent session DTOs + invoke wrappers

冲突解决:
- migrations/mod.rs: upstream v133_create_session_events → v139，保留本地 v133-v138 OPC 系列
- 其余 200+ 文件自动合并（i18n 11 语言 + Cargo.lock + package.json）
- 保留本地: AxInvest 描述 / oxlint 工具链 / plugin-updater ^2.11.0 / OPC+股票 16 crate

+++ src-tauri/crates/runtime/src/persistent_queue.rs
@@ -49,19 +49,14 @@ impl Default for Model {
 }
 
 fn now_ts() -> i64 {
-    std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_secs() as i64
+    std::time::SystemTime::now().duration_since(std::time::UNIX_EPOCH).unwrap_or_default().as_secs()
+        as i64
 }
 
 fn uuid_v4() -> String {
     use std::time::{SystemTime, UNIX_EPOCH};
-    let timestamp = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_nanos();
-    let random: u128 = (timestamp as u128) << 64 | (rand_u64() as u128);
+    let timestamp = SystemTime::now().duration_since(UNIX_EPOCH).unwrap_or_default().as_nanos();
+    let random: u128 = timestamp << 64 | (rand_u64() as u128);
     format!("{:032x}", random)
 }
